move `_bootstrap_components` to root node. closes #771
