+++ bsb/config/_make.py
@@ -9,6 +9,7 @@ from re import sub
 import errr
 
 from ..exceptions import (
+    BootError,
     CastError,
     ConfigurationError,
     DynamicClassError,
@@ -309,9 +310,16 @@ def compile_postnew(cls):
 
 
 def wrap_root_postnew(post_new):
-    def __post_new__(self, *args, _parent=None, _key=None, **kwargs):
+    def __post_new__(self, *args, _parent=None, _key=None, _store=None, **kwargs):
         if not hasattr(self, "_meta"):
             self._meta = {"path": None, "produced": True}
+
+        # Root node bootstrapping sequence
+        try:
+            _bootstrap_components(kwargs.get("components", []), file_store=_store)
+        except Exception as e:
+            raise BootError("Failed to bootstrap configuration.") from e
+
         try:
             with warnings.catch_warnings(record=True) as log:
                 try:
@@ -348,6 +356,15 @@ def _bubble_up_warnings(log):
             warn(str(m) + " in " + m.node.get_node_name() + attr, type(m), stacklevel=4)
         else:
             warn(str(m), w.category, stacklevel=4)
+
+
+def _bootstrap_components(components, file_store=None):
+    from bsb.storage._files import CodeDependencyNode
+
+    for component in components:
+        component_node = CodeDependencyNode(component)
+        component_node.file_store = file_store
+        component_node.load_object()
 
 
 def get_config_attributes(cls):
